Update for faster PBA loading

+++ linux/DtaDevLinuxDrive.h
@@ -42,7 +42,7 @@ public:
      * @param bufferlen length of the input/output buffer
      */
     virtual uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-            void * buffer, uint16_t bufferlen) = 0;
+            void * buffer, uint32_t bufferlen) = 0;
     /** Routine to send an identify to the device */
     virtual void identify(OPAL_DiskInfo& disk_info) = 0;
 };
