+++ linux/DtaDevOS.cpp
@@ -83,7 +83,7 @@ void DtaDevOS::init(const char * devref)
 }
 
 uint8_t DtaDevOS::sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-	void * buffer, uint16_t bufferlen)
+	void * buffer, uint32_t bufferlen)
 {
 	if (!isOpen) return 0xfe; //disk open failed so this will too
